mas bima ini tolong dianuin

+++ database/migrations/0001_01_01_000002_create_jobs_table.php
@@ -24,14 +24,14 @@ return new class extends Migration
         Schema::create('job_batches', function (Blueprint $table) {
             $table->string('id')->primary();
             $table->string('name');
-            $table->integer('total_jobs');
-            $table->integer('Pending_jobs');
-            $table->integer('failed_jobs');
+            $table->bigInteger('total_jobs');
+            $table->bigInteger('Pending_jobs');
+            $table->bigInteger('failed_jobs');
             $table->longText('failed_job_ids');
             $table->mediumText('options')->nullable();
-            $table->integer('cancelled_at')->nullable();
-            $table->integer('created_at');
-            $table->integer('finished_at')->nullable();
+            $table->bigInteger('cancelled_at')->nullable();
+            $table->bigInteger('created_at');
+            $table->bigInteger('finished_at')->nullable();
         });
 
         Schema::create('failed_jobs', function (Blueprint $table) {
