+++ database/migrations/2026_07_06_220007_create_cuti_izins_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
             $table->string('jenis_cuti_izin');               // Jenis Cuti/Izin (Cuti Tahunan, Sakit, dll)
             $table->date('tanggal_mulai');                   // Tanggal Mulai
             $table->date('tanggal_selesai');                 // Tanggal Selesai
-            $table->integer('lama_hari');                    // Lama Hari
+            $table->bigInteger('lama_hari');                    // Lama Hari
             $table->text('alasan');                          // Alasan
             $table->string('status');                        // Status (Disetujui, Ditolak, Pending)
             $table->timestamps();
