+++ database/migrations/2026_07_08_200005_create_vendor_performances_table.php
@@ -11,10 +11,10 @@ return new class extends Migration
         Schema::create('vendor_performances', function (Blueprint $table) {
             $table->id();
             $table->string('vendor');
-            $table->integer('total_order')->default(0);
+            $table->bigInteger('total_order')->default(0);
             $table->decimal('ketepatan_waktu', 5, 2)->default(0)->comment('persen');
             $table->decimal('kualitas_barang', 5, 2)->default(0)->comment('skala 1-100');
-            $table->integer('komplain')->default(0);
+            $table->bigInteger('komplain')->default(0);
             $table->decimal('penilaian_akhir', 5, 2)->default(0)->comment('skala 1-100');
             $table->text('catatan')->nullable();
             $table->timestamps();
